Fixed issue in plots.py and added two 1D models
In detail, the AnchoredVTexts had a bug due to an update in
matplotlib offsetbox. Removing minimumdescent=True fixed the issue.

In custom_astropy_models.py I added a piecewise linear 1d function and
a "Split Gaussian", which is just another version of the asymmetric
gaussian but consisting in two gaussians of different widths stitched
at the center.